+++ src/index.js
@@ -1,4 +1,4 @@
 import { MicroscopyViewerLayer } from './layers';
 import { MicroscopyViewer } from './microscopy-viewer';
 
-export { MicroscopyViewerLayer, MicroscopyViewer };
+export { MicroscopyViewerLayer as ViewerLayer, MicroscopyViewer as Viewer };